requirements.txt
flask>=3.0.0
flask-sqlalchemy>=3.1.0
flask-cors>=4.0.0
flask-apscheduler>=1.13.0
pymysql>=1.1.0
cryptography>=42.0.0
langgraph>=0.2.0
pydantic>=2.6.0
python-dotenv>=1.0.0
litellm>=1.40.0
redis>=5.0.0
mcp>=1.0.0
httpx>=0.27.0
tenacity>=8.2.3
